# docs/swagger.yaml
basePath: /
definitions:
...
  controller.cambiarPasswordRequest:
    properties:
      password:
        type: string
    type: object
...
  controller.empresaResponse:
    properties:
      creado_en:
        type: string
      estado:
        type: boolean
      id:
        type: integer
      maximo_usuarios:
        type: integer
      moneda:
        type: string
      nombre:
        type: string
      pais:
        type: string
      vencimiento:
        type: string
    type: object
  controller.errorResponse:
    properties:
      message:
        type: string
    type: object
...
  controller.usuarioLoginRequest:
    properties:
      contrasena:
        type: string
      usuario:
        type: string
    type: object
  controller.usuarioLoginResponse:
    properties:
      empresa:
        $ref: '#/definitions/controller.empresaResponse'
      empresa_id:
        type: integer
      token:
        type: string
      user:
        $ref: '#/definitions/controller.usuarioResponse'
    type: object
  controller.usuarioResponse:
    properties:
      empresa_id:
        type: integer
      id:
        type: integer
      usuario:
        type: string
    type: object
...
info:
  contact: {}
  description: API para autenticacion, administracion de empresas y catalogos de soporte
    para el frontend de Rentals Go.
  title: Rentals Go API
  version: "1.0"
paths:
...
  /api/me/password:
    patch:
      consumes:
      - application/json
      description: Permite al usuario autenticado cambiar su propia contraseña.
      parameters:
      - description: Nueva contraseña
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/controller.cambiarPasswordRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: No autorizado
          schema:
            $ref: '#/definitions/controller.errorResponse'
      security:
      - ApiKeyAuth: []
      summary: Cambiar contraseña del usuario actual
      tags:
      - Usuarios
...
  /api/user/servicios/{id}:
    delete:
      responses: {}
      summary: Eliminar medición
      tags:
      - Servicios
    get:
      responses: {}
      summary: Obtener detalle de medición
      tags:
      - Servicios
    put:
      parameters:
      - description: ID de la medición
        in: path
        name: id
        required: true
        type: integer
      - description: Nueva lectura_actual
        in: body
        name: request
        required: true
        schema:
          additionalProperties:
            format: float64
            type: number
          type: object
      responses: {}
      summary: Editar lectura errónea
      tags:
      - Servicios
  /api/user/staff:
    get:
      parameters:
      - description: ID de la empresa
        in: query
        name: empresa_id
        required: true
        type: integer
      - description: Página
        in: query
        name: pag
        type: integer
      - description: Items por página
        in: query
        name: por_pagina
        type: integer
      - description: Buscar por nombre de usuario
        in: query
        name: buscar
        type: string
      responses: {}
      summary: Listar staff de la empresa
      tags:
      - Staff
    post:
      responses: {}
      summary: Registrar nuevo empleado
      tags:
      - Staff
...
  /auth/login:
    post:
      consumes:
      - application/json
      description: Permite a un usuario autenticarse y obtener un token JWT. Además,
        establece una cookie de sesión.
      parameters:
      - description: Credenciales de usuario
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/controller.usuarioLoginRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/controller.usuarioLoginResponse'
        "401":
          description: Credenciales inválidas
          schema:
            $ref: '#/definitions/controller.errorResponse'
      summary: Iniciar sesión como usuario
      tags:
      - Usuarios
...
  /me:
    get:
      description: Retorna los datos del usuario autenticado y su empresa.
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/controller.usuarioLoginResponse'
        "401":
          description: No autorizado
          schema:
            $ref: '#/definitions/controller.errorResponse'
      security:
      - ApiKeyAuth: []
      summary: Obtener perfil del usuario actual
      tags:
      - Usuarios
schemes:
- http
- https
securityDefinitions:
  BearerAuth:
    description: 'Token JWT en formato: Bearer <token>'
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"